Adding my resume to my github pages.
This commit also includes a script I used to generate the pdf versions of my resume as well as the unredacted versions. The main, checked in version of the resume is in html and, for privacy reasons, does not include my phone or email, but these are added into a private version of the resume with the script.